Remove H3Fat from geoToH3
